Remove previous implementation
refs hitobito/hitobito_sac_cas#255
