+++ app/models/youth/person.rb
@@ -23,9 +23,6 @@ module Youth::Person
     has_many :managers, through: :people_managers
     has_many :manageds, through: :people_manageds
 
-    accepts_nested_attributes_for :people_managers, allow_destroy: true
-    accepts_nested_attributes_for :people_manageds, allow_destroy: true
-
     validates :nationality_j_s, inclusion: { in: NATIONALITIES_J_S, allow_blank: true }
 
     validate :assert_either_only_managers_or_manageds
