--- app/views/people/manageds/_create_new_managed_person.html.haml
@@ -1,10 +1,0 @@
-= f.fields_for(:managed_attributes, Person.new) do |fields|
-  = field_set_tag do
-    = fields.labeled_input_fields :first_name, :last_name, required: true
-    = fields.labeled(:gender) do
-      - (Person::GENDERS + ['']).each do |key|
-        = fields.inline_radio_button(:gender, key, entry.gender_label(key))
-  
-    = fields.labeled_input_field(:birthday,
-      help_inline: t('.birthday_hint'),
-      class: 'col-2 form-control form-control-sm d-inline')
