--- app/views/people/manageds/_fields.html.haml
@@ -1,15 +1,0 @@
--#  Copyright (c) 2023, CEVI Schweiz, Pfadibewegung Schweiz,
--#  Jungwacht Blauring Schweiz, Pro Natura, Stiftung für junge Auslandschweizer.
--#  This file is part of hitobito_youth and licensed under the Affero General Public License version 3
--#  or later. See the COPYING file at the top-level directory or at
--#  https://github.com/hitobito/hitobito_youth.
-
-- if f.object.new_record?
-  - if cannot?(:lookup_manageds, Person) && FeatureGate.enabled?('people.people_managers.self_service_managed_creation')
-    = render('people/manageds/create_new_managed_person', f: f)
-  - else
-    %span
-      = f.person_field(:managed, { disabled: cannot?(:create_households, Person), data: { url: query_people_path(limit_by_permission: :change_managers) } })
-      %span.help-block= t('.manageds_help')
-- else
-  = f.object.managed.full_name
