--- app/views/people/managers/_fields.html.haml
@@ -1,11 +1,0 @@
--#  Copyright (c) 2023, CEVI Schweiz, Pfadibewegung Schweiz,
--#  Jungwacht Blauring Schweiz, Pro Natura, Stiftung für junge Auslandschweizer.
--#  This file is part of hitobito_youth and licensed under the Affero General Public License version 3
--#  or later. See the COPYING file at the top-level directory or at
--#  https://github.com/hitobito/hitobito_youth.
-
-- if f.object.new_record?
-  = f.person_field(:manager)
-- else
-  .col-4
-    = f.object.manager.full_name
